ISSUE #58 - Auto-Minification in postinstall
- add minifier in package.json
- The postinstall script will now automatically minifies all scripts
that is required for external use (formerly listed in config.js as
external) and put them in the relevant folder.

- external is removed from config.js and .min.(js|css) or .(js|css) is
determined by logfile.js_debug_level

- a common_public_files.js is generated by postinstall that will list
all the files that was used to be in config.external, and it is hooked
up onto js/core/config.js. the change should technically be seamless

- updated config_sample.js to reflect this change.